requirements.txt
streamlit==1.32.0
plotly==5.18.0
pyvista==0.43.0
numpy==1.26.4
pillow==10.2.0
pandas==2.2.1 